requirements.txt
et_xmlfile==2.0.0
numpy==2.4.4
openpyxl==3.1.5
pandas==3.0.2
psycopg2-binary==2.9.12
PySide6==6.11.0
PySide6_Addons==6.11.0
PySide6_Essentials==6.11.0
python-dateutil==2.9.0.post0
shiboken6==6.11.0
six==1.17.0
